Navigation Drawer using MVP
