+++ app/build.gradle
@@ -55,7 +55,6 @@ dependencies {
     implementation 'com.google.android.material:material:1.0.0'
     implementation "androidx.navigation:navigation-fragment:$nav_version"
     implementation "androidx.navigation:navigation-ui:$nav_version"
-    implementation "androidx.lifecycle:lifecycle-extensions:$nav_version"
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
