+++ app/src/main/java/com/developers_recipes/recetario/contracts/ApiContract.kt
@@ -14,5 +14,6 @@ interface ApiContract {
     interface Presenter{
         fun showApiVersion(version: String)
         fun setApiVersion(version: String)
+        fun fetchApiVersion()
     }
 }
