+++ app/src/main/java/com/developers_recipes/recetario/ui/explore/ExploreFragment.kt
@@ -4,28 +4,18 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
 import com.developers_recipes.recetario.R
 
 class ExploreFragment : Fragment() {
-
-    private lateinit var exploreViewModel: ExploreViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        exploreViewModel =
-            ViewModelProviders.of(this).get(ExploreViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_explore, container, false)
-        val textView: TextView = root.findViewById(R.id.text_slideshow)
-        exploreViewModel.text.observe(this, Observer {
-            textView.text = it
-        })
+
         return root
     }
 }
