--- app/src/main/java/com/developers_recipes/recetario/ui/explore/ExploreViewModel.kt
@@ -1,13 +1,0 @@
-package com.developers_recipes.recetario.ui.explore
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ExploreViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is explore Fragment"
-    }
-    val text: LiveData<String> = _text
-}
