--- app/src/main/java/com/developers_recipes/recetario/ui/profile/ProfileViewModel.kt
@@ -1,13 +1,0 @@
-package com.developers_recipes.recetario.ui.profile
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ProfileViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is profile Fragment"
-    }
-    val text: LiveData<String> = _text
-}
