+++ app/src/main/java/com/developers_recipes/recetario/ui/recipe/RecipeFragment.kt
@@ -4,28 +4,17 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
 import com.developers_recipes.recetario.R
 
 class RecipeFragment : Fragment() {
-
-    private lateinit var recipeViewModel: RecipeViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        recipeViewModel =
-            ViewModelProviders.of(this).get(RecipeViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_recipe, container, false)
-        val textView: TextView = root.findViewById(R.id.text_tools)
-        recipeViewModel.text.observe(this, Observer {
-            textView.text = it
-        })
         return root
     }
 }
