--- app/src/main/java/com/developers_recipes/recetario/ui/recipe/RecipeViewModel.kt
@@ -1,13 +1,0 @@
-package com.developers_recipes.recetario.ui.recipe
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class RecipeViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is recipe Fragment"
-    }
-    val text: LiveData<String> = _text
-}
