+++ app/src/main/res/values/strings.xml
@@ -6,15 +6,21 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
 
-    <string name="navigation_drawer_open">Abrir menú</string>
-    <string name="navigation_drawer_close">Cerrar menú</string>
-    <string name="nav_header_title">Nombre Usuario</string>
-    <string name="nav_header_subtitle">usuario@dominio.com</string>
-    <string name="nav_header_desc">Cabecera</string>
-    <string name="action_settings">Ajustes</string>
+    <string name="navigation_drawer_open">Open menu</string>
+    <string name="navigation_drawer_close">Close menu</string>
+    <string name="nav_header_title">Username</string>
+    <string name="nav_header_subtitle">user@domain.com</string>
+    <string name="nav_header_desc">Header</string>
+    <string name="action_settings">Settings</string>
 
     <string name="menu_home">Home</string>
     <string name="menu_profile">Profile</string>
     <string name="menu_explore">Explore</string>
     <string name="menu_recipe">Recipe</string>
+
+
+    <string name="explore_fragment_help">This is explore Fragment</string>
+    <string name="home_fragment_help">This is home Fragment</string>
+    <string name="profile_fragment_help">This is profile Fragment</string>
+    <string name="recipe_fragment_help">This is recipe Fragment</string>
 </resources>
